add,delete,set alarm for tasks

+++ app/src/main/java/com/raywenderlich/tasksapp/data/Note.kt
@@ -3,13 +3,14 @@ package com.raywenderlich.tasksapp.data
 import android.os.Parcelable
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import com.raywenderlich.tasksapp.tools.IDGenerator
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
 @Entity(tableName = "note_table")
 data class Note(
     @PrimaryKey(autoGenerate = true)
-    val id: Int,
+    val id: Long,
     val title: String,
     val description: String,
     val date: String,
